+++ substrate/bin/node/cli/Cargo.toml
@@ -90,6 +90,7 @@ sc-storage-monitor = { path = "../../../client/storage-monitor" }
 sc-offchain = { path = "../../../client/offchain" }
 
 # frame dependencies
+frame-benchmarking = { path = "../../../frame/benchmarking" }
 frame-system = { path = "../../../frame/system" }
 frame-system-rpc-runtime-api = { path = "../../../frame/system/rpc/runtime-api" }
 pallet-assets = { path = "../../../frame/assets" }
@@ -102,7 +103,6 @@ pallet-skip-feeless-payment = { path = "../../../frame/transaction-payment/skip-
 kitchensink-runtime = { path = "../runtime" }
 node-rpc = { path = "../rpc" }
 node-primitives = { path = "../primitives" }
-node-executor = { package = "staging-node-executor", path = "../executor" }
 
 # CLI-specific dependencies
 sc-cli = { path = "../../../client/cli", optional = true}
@@ -136,6 +136,26 @@ substrate-rpc-client = { path = "../../../utils/frame/rpc/client" }
 pallet-timestamp = { path = "../../../frame/timestamp" }
 substrate-cli-test-utils = { path = "../../../test-utils/cli" }
 
+wat = "1.0"
+frame-support = { path = "../../../frame/support" }
+node-testing = { path = "../testing" }
+pallet-balances = { path = "../../../frame/balances" }
+pallet-contracts = { path = "../../../frame/contracts" }
+pallet-glutton = { path = "../../../frame/glutton" }
+pallet-sudo = { path = "../../../frame/sudo" }
+pallet-treasury = { path = "../../../frame/treasury" }
+pallet-transaction-payment = { path = "../../../frame/transaction-payment" }
+sp-application-crypto = { path = "../../../primitives/application-crypto" }
+pallet-root-testing = { path = "../../../frame/root-testing" }
+sp-consensus-babe = { path = "../../../primitives/consensus/babe" }
+sp-externalities = { path = "../../../primitives/externalities" }
+sp-keyring = { path = "../../../primitives/keyring" }
+sp-runtime = { path = "../../../primitives/runtime" }
+serde_json = "1.0.108"
+scale-info = { version = "2.10.0", features = ["derive", "serde"] }
+sp-trie = { path = "../../../primitives/trie" }
+sp-state-machine = { path = "../../../primitives/state-machine" }
+
 [build-dependencies]
 clap = { version = "4.4.6", optional = true }
 clap_complete = { version = "4.0.2", optional = true }
@@ -163,16 +183,21 @@ cli = [
 ]
 runtime-benchmarks = [
 	"frame-benchmarking-cli/runtime-benchmarks",
+	"frame-benchmarking/runtime-benchmarks",
+	"frame-support/runtime-benchmarks",
 	"frame-system/runtime-benchmarks",
 	"kitchensink-runtime/runtime-benchmarks",
-	"node-executor/runtime-benchmarks",
 	"node-inspect?/runtime-benchmarks",
 	"pallet-asset-tx-payment/runtime-benchmarks",
 	"pallet-assets/runtime-benchmarks",
 	"pallet-balances/runtime-benchmarks",
+	"pallet-contracts/runtime-benchmarks",
+	"pallet-glutton/runtime-benchmarks",
 	"pallet-im-online/runtime-benchmarks",
 	"pallet-skip-feeless-payment/runtime-benchmarks",
+	"pallet-sudo/runtime-benchmarks",
 	"pallet-timestamp/runtime-benchmarks",
+	"pallet-treasury/runtime-benchmarks",
 	"sc-client-db/runtime-benchmarks",
 	"sc-service/runtime-benchmarks",
 	"sp-runtime/runtime-benchmarks",
@@ -180,15 +205,22 @@ runtime-benchmarks = [
 # Enable features that allow the runtime to be tried and debugged. Name might be subject to change
 # in the near future.
 try-runtime = [
+	"frame-support/try-runtime",
 	"frame-system/try-runtime",
 	"kitchensink-runtime/try-runtime",
 	"pallet-asset-conversion-tx-payment/try-runtime",
 	"pallet-asset-tx-payment/try-runtime",
 	"pallet-assets/try-runtime",
 	"pallet-balances/try-runtime",
+	"pallet-contracts/try-runtime",
+	"pallet-glutton/try-runtime",
 	"pallet-im-online/try-runtime",
+	"pallet-root-testing/try-runtime",
 	"pallet-skip-feeless-payment/try-runtime",
+	"pallet-sudo/try-runtime",
 	"pallet-timestamp/try-runtime",
+	"pallet-transaction-payment/try-runtime",
+	"pallet-treasury/try-runtime",
 	"sp-runtime/try-runtime",
 	"substrate-cli-test-utils/try-runtime",
 	"try-runtime-cli/try-runtime",
@@ -200,4 +232,8 @@ harness = false
 
 [[bench]]
 name = "block_production"
+harness = false
+
+[[bench]]
+name = "executor"
 harness = false
