+++ substrate/bin/node/cli/tests/res/default_genesis_config.json
