--- substrate/bin/node/executor/Cargo.toml
@@ -1,72 +1,0 @@
-[package]
-name = "staging-node-executor"
-version = "3.0.0-dev"
-authors.workspace = true
-description = "Substrate node implementation in Rust."
-edition.workspace = true
-license = "Apache-2.0"
-homepage = "https://substrate.io"
-repository.workspace = true
-publish = false
-
-[package.metadata.docs.rs]
-targets = ["x86_64-unknown-linux-gnu"]
-
-[dependencies]
-codec = { package = "parity-scale-codec", version = "3.6.1" }
-scale-info = { version = "2.10.0", features = ["derive", "serde"] }
-frame-benchmarking = { path = "../../../frame/benchmarking" }
-node-primitives = { path = "../primitives" }
-kitchensink-runtime = { path = "../runtime" }
-sc-executor = { path = "../../../client/executor" }
-sp-core = { path = "../../../primitives/core", features=["serde"] }
-sp-io = { path = "../../../primitives/io" }
-sp-keystore = { path = "../../../primitives/keystore" }
-sp-state-machine = { path = "../../../primitives/state-machine" }
-sp-tracing = { path = "../../../primitives/tracing" }
-sp-trie = { path = "../../../primitives/trie" }
-sp-statement-store = { path = "../../../primitives/statement-store", features=["serde"] }
-
-[dev-dependencies]
-criterion = "0.4.0"
-futures = "0.3.21"
-wat = "1.0"
-frame-support = { path = "../../../frame/support" }
-frame-system = { path = "../../../frame/system" }
-node-testing = { path = "../testing" }
-pallet-balances = { path = "../../../frame/balances" }
-pallet-contracts = { path = "../../../frame/contracts" }
-pallet-im-online = { path = "../../../frame/im-online" }
-pallet-glutton = { path = "../../../frame/glutton" }
-pallet-sudo = { path = "../../../frame/sudo" }
-pallet-timestamp = { path = "../../../frame/timestamp" }
-pallet-treasury = { path = "../../../frame/treasury" }
-pallet-transaction-payment = { path = "../../../frame/transaction-payment" }
-sp-application-crypto = { path = "../../../primitives/application-crypto" }
-pallet-root-testing = { path = "../../../frame/root-testing" }
-sp-consensus-babe = { path = "../../../primitives/consensus/babe" }
-sp-externalities = { path = "../../../primitives/externalities" }
-sp-keyring = { path = "../../../primitives/keyring" }
-sp-runtime = { path = "../../../primitives/runtime" }
-serde_json = "1.0.108"
-
-[features]
-stress-test = []
-runtime-benchmarks = [
-	"frame-benchmarking/runtime-benchmarks",
-	"frame-support/runtime-benchmarks",
-	"frame-system/runtime-benchmarks",
-	"kitchensink-runtime/runtime-benchmarks",
-	"pallet-balances/runtime-benchmarks",
-	"pallet-contracts/runtime-benchmarks",
-	"pallet-glutton/runtime-benchmarks",
-	"pallet-im-online/runtime-benchmarks",
-	"pallet-sudo/runtime-benchmarks",
-	"pallet-timestamp/runtime-benchmarks",
-	"pallet-treasury/runtime-benchmarks",
-	"sp-runtime/runtime-benchmarks",
-]
-
-[[bench]]
-name = "bench"
-harness = false
