+++ substrate/bin/node/inspect/Cargo.toml
@@ -15,17 +15,17 @@ targets = ["x86_64-unknown-linux-gnu"]
 clap = { version = "4.4.6", features = ["derive"] }
 codec = { package = "parity-scale-codec", version = "3.6.1" }
 thiserror = "1.0"
-node-executor = { package = "staging-node-executor", path = "../executor" }
 sc-cli = { path = "../../../client/cli" }
 sc-client-api = { path = "../../../client/api" }
 sc-service = { path = "../../../client/service", default-features = false}
 sp-blockchain = { path = "../../../primitives/blockchain" }
 sp-core = { path = "../../../primitives/core" }
+sp-io = { path = "../../../primitives/io" }
 sp-runtime = { path = "../../../primitives/runtime" }
+sp-statement-store = { path = "../../../primitives/statement-store" }
 
 [features]
 runtime-benchmarks = [
-	"node-executor/runtime-benchmarks",
 	"sc-service/runtime-benchmarks",
 	"sp-runtime/runtime-benchmarks",
 ]
